code polishing
